% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{simulator}
\alias{simulator}
\title{Simulation setup}
\usage{
simulator(
  rule,
  criterion,
  df_param,
  ntrials = 10,
  specific_trials = NA,
  cores = 1,
  shuffle_group = NA,
  chunking_num = nrow(df_param),
  required_packages = NULL,
  filepath = NA,
  verbose = T
)
}
\arguments{
\item{rule}{function}

\item{criterion}{function}

\item{df_param}{data frame}

\item{ntrials}{number of trials for each row}

\item{specific_trials}{vector of integers of specific trials to run}

\item{cores}{number of cores, where if larger than \code{1}, parallelization occurs}

\item{shuffle_group}{either \code{NA} or a list where each element of the list
is a vector containing unique integers ranging within \code{1:nrow(df_param)}}

\item{chunking_num}{integer, where if \code{!is.na(chunking_num)}, dictatees how many
intermediary files are saved during the simulation}

\item{required_packages}{packages to load into the multisession if
\code{cores>1}}

\item{filepath}{string}

\item{verbose}{boolean}
}
\value{
list
}
\description{
This function applies \code{rule} (which takes in parameter settings and outputs
a synthetic dataset) and \code{criterion} (which takes a synthetic dataset and outputs
the results from estimators) to all the rows of \code{df_param} (which is a matrix
that contains different parameter settings for each row). The distinction between
\code{rule} and \code{criterion} is only made by the user, as the user can design
exactly the same simulation that uses one but not the other.
}
\details{
The input to \code{rule} must be a vector (a row from \code{df_param}), while
the input to \code{criterion} must be first the output of \code{rule} and second
a vector (the same row from \code{df_param}). Both these functions is allowed
to output lists.

The output to \code{simulator} is a list, one element for each
row of \code{df_param}. Each element of the list is typically a list or
a matrix. This depends on how the user set up what \code{criterion} returns.

The function has a \code{tryCatch} call, so if an error happens, the result for that
trial and row of \code{df_param} will be an \code{NA}.

The remaining inputs for \code{simulator} are cosmetic.

\code{filepath} is a filepath to the temporary save location. If set to not \code{NA},
\code{simulator} will save the results of every row of \code{df_param} there
as it runs the simulations.
}
